x-original-posts/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
tweepy>=4.14.0
